Update BetaWithNoTime
Made some minor changes, tried to add the timer,  but it significantly slowed down the speed of the game. So I scrapped it. But I commented through all of the program. It still has the problem of not updating the right side of the GUI when all of the letters have been collected, but it WILL allow you to type the password and end the game if you know it.